requirements.txt
pypinyin
tiktoken
socksio
tqdm
colorama
Pygments
langchain
openai
langchain_experimental
langchain_openai
markdown
PyPDF2
pdfplumber
pandas
commentjson
openpyxl
pandoc
wolframalpha
faiss-cpu==1.7.4
duckduckgo-search>=3.8
unstructured[all-docs]
tabulate
ujson
python-docx
websocket_client
pydantic==1.10.8
olefile
python-pptx
fastapi
uvicorn
python-multipart
sentence_transformers
langchain-community
rank-bm25
chromadb
apscheduler
asyncio
requests
aiokafka
pymysql
aiomysql
xlrd
sqlalchemy
python-jose
passlib
mysqlclient
bcrypt
psutil
pillow
lxml
matplotlib
opencv-python
pdf2image
pytesseract
streamlit